Refactor driver to rider
