+++ app/src/main/AndroidManifest.xml
@@ -38,11 +38,11 @@
         </activity>
 
         <activity
-            android:name=".fragments.DriverMainMap"
+            android:name=".fragments.RiderMainMap"
             android:label="@string/title_activity_driver_main_map"></activity>
             
         <activity
-            android:name=".activities.DriverMainActivity"
+            android:name=".activities.RiderMainActivity"
             android:label="@string/title_activity_driver_main"
             android:theme="@style/AppTheme.NoActionBar" />
     </application>
