+++ app/src/main/res/layout/activity_driver_main_map.xml
@@ -6,4 +6,4 @@ android:id="@+id/map"
 android:name="com.google.android.gms.maps.SupportMapFragment"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".fragments.DriverMainMap" />
+tools:context=".fragments.RiderMainMap" />
