+++ app/src/main/res/layout/content_driver_main.xml
@@ -25,7 +25,7 @@
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".fragments.DriverMainMap"
+        tools:context=".fragments.RiderMainMap"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toTopOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
